Add tests for  store
